Done backend customer-list

+++ spring-boot-web/src/main/java/com/javaweb/service/CustomerService.java
@@ -13,5 +13,5 @@ public interface CustomerService {
     List<CustomerDTO> findAll(CustomerSearchRequest params, Pageable pageable);
     int countTotalItems(CustomerSearchRequest params);
     boolean isStaffOfCustomer(Long staffId, Long customerId);
-    CustomerDTO findByIdAndIsActive(Long id, Integer isActive);
+    CustomerDTO findByIdAndIsActive(Long id, Long isActive);
 }
